Build fix.
svn:sync-xref-src-repo-rev: r123619

+++ src/VBox/Devices/Audio/DrvAudioCommon.cpp
@@ -1,4 +1,4 @@
-/* $Id: DrvAudioCommon.cpp 73058 2018-07-11 10:12:55Z andreas.loeffler@oracle.com $ */
+/* $Id: DrvAudioCommon.cpp 73059 2018-07-11 10:16:05Z andreas.loeffler@oracle.com $ */
 /** @file
  * Intermedia audio driver, common routines.
  *
@@ -1121,7 +1121,7 @@ uint32_t DrvAudioHlpMsToFrames(const PPDMAUDIOPCMPROPS pProps, uint32_t uMs)
 {
     AssertPtrReturn(pProps, 0);
 
-    const size_t cbFrame = (pProps->cBits / 8) * pProps->cChannels;
+    const uint32_t cbFrame = (pProps->cBits / 8) * pProps->cChannels;
     if (!cbFrame) /* Prevent division by zero. */
         return 0;
